refactor: Add zone-options-dropdown
