+++ src/app/zone/zone-list/zone-list.component.html
@@ -7,19 +7,18 @@
     [routerLink]="['/audits', zone.auditId, 'zones', zone.zoneId]"
     [percentageQuery]="{progressType: 'zone', auditId: zone.auditId, zoneId: zone.zoneId}"
   >
-    <app-option-dropdown options style="z-index: 900">
-      <button ngbDropdownItem class="bi-pencil-square" (click)="rename(zone)">
-        Rename
-      </button>
-      <button ngbDropdownItem class="bi-copy" (click)="duplicate(zone)">
-        Duplicate
-      </button>
-      <ng-container>
-        <div class="dropdown-divider"></div>
-        <button ngbDropdownItem class="text-danger bi-trash" (click)="delete(zone)">
-          Delete
+    <div options style="z-index: 900" (click)="$event.stopImmediatePropagation(); $event.preventDefault()">
+      <app-zone-options-dropdown [zone]="zone">
+        <button ngbDropdownItem class="bi-copy" (click)="duplicate(zone)">
+          Duplicate
         </button>
-      </ng-container>
-    </app-option-dropdown>
+        <ng-container>
+          <div class="dropdown-divider"></div>
+          <button ngbDropdownItem class="text-danger bi-trash" (click)="delete(zone)">
+            Delete
+          </button>
+        </ng-container>
+      </app-zone-options-dropdown>
+    </div>
   </app-feature-card>
 }
