+++ src/app/zone/zone-list/zone-list.component.ts
@@ -1,9 +1,10 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
-import {switchMap} from 'rxjs';
-import {AuditZoneService} from 'src/app/shared/services/audit-zone.service';
-import {Zone} from '../../shared/model/zone.interface';
 import {ToastService} from '@mean-stream/ngbx';
+import {switchMap} from 'rxjs';
+
+import {Zone} from '../../shared/model/zone.interface';
+import {AuditZoneService} from '../../shared/services/audit-zone.service';
 
 @Component({
   selector: 'app-zone-list',
@@ -42,21 +43,6 @@ export class ZoneListComponent implements OnInit {
     }).subscribe(({data}) => {
       this.zones.push(data);
       this.toastService.success('Create Zone', 'Successfully created new zone.');
-    });
-  }
-
-  rename(zone: Zone) {
-    const name = prompt('Rename Zone ', zone.zoneName);
-    if (!name) {
-      return;
-    }
-    this.zoneService.updateAuditZone(zone.auditId, zone.zoneId, {
-      ...zone, zoneName: name,
-    }).subscribe(({data}) => {
-      const index = this.zones.indexOf(zone);
-      this.zones[index] = data;
-      this.toastService.success('Rename Zone', 'Successfully renamed zone.');
-      this.toastService.success('Rename Zone', 'Successfully renamed zone.');
     });
   }
