+++ src/app/zone/zone-options-dropdown/zone-options-dropdown.component.html
@@ -1,0 +1,10 @@
+<div ngbDropdown class="d-inline-block">
+  <button type="button" class="btn btn-link bi-three-dots" id="dropdown-audit" ngbDropdownToggle>
+  </button>
+  <div ngbDropdownMenu aria-labelledby="dropdown-audit">
+    <button ngbDropdownItem class="bi-pencil-square" (click)="rename()">
+      Rename
+    </button>
+    <ng-content></ng-content>
+  </div>
+</div>
